src/flask/requirements.txt
pyjwt
bcrypt
mysql-connector
flask
flask_cors
flask_restplus
Pillow
